refactor create key value store methods
